Adding method to print all drugs, fixing case when deleting patient, now removes all his doses and drugs aswell with recovering them to the pool

+++ Szpital/src/main/webapp/databaseHelpers/Patient.java
@@ -21,9 +21,13 @@ public class Patient {
 
 		boolean result;
 
+		result = Drug.removeAllDrugFromPatient(patientID);
 		String SQL = "DELETE FROM Pacjent WHERE Pacjent_ID = " + patientID;
+		
+		if(result){
 		result = Database.deleteFromDatabase(SQL);
-
+		}
+		
 		if (result) {
 			result = updateAmountOfPatients(hospitalWard_ID, -1);
 		}
